feat: Add pandas library and support for CSV files
Add *.csv to the .gitignore file to exclude CSV files from version
control.
In the Pipfile, add pandas as a package dependency. This library allows
working with data in CSV format. The addition of pandas enables the use
of powerful data manipulation and analysis capabilities. This change
also
includes updates to the Pipfile.lock file. The main.py file is modified
to
import the SetInput and TerminalOutput classes. Moreover, the SetInput
class
is updated to include the creation of a SetInput instance and setting
the font
for the input text field. Finally, the SystemData class in the
system_data.py
file is modified to correct the file paths for loading and saving JSON
files.
This change improves the functionality of the application by enabling
the
manipulation and analysis of CSV data. No issues are associated with
this
commit.