feat: Enable saving data to JSON files and unit JSON file
This code change adds functionality that allows data to be saved to JSON
files and a unit JSON file. The "data.save_to_json_files()" and
"data.save_to_unit_json_file()" methods have been called in the
"start_callback" function in the main.py file and the
"SetInput.start_trade_instance" method in the set_input.py file. This
change enhances data persistence and facilitates easy retrieval and
usage of the saved data. The addition of this capability improves the
overall user experience and provides better support for analytics and
reporting.